feat(map): extract add information flow to its own component
