+++ map/src/components/add-instructions.tsx
@@ -2,24 +2,133 @@ import React from 'react';
 import { MdClose } from 'react-icons/md';
 
 import styled from '../styling';
+import { button, buttonPrimary } from '../styling/mixins';
+
+interface AddInfo {
+  marker: google.maps.Marker | null;
+  circle: google.maps.Circle | null;
+}
 
 interface Props {
   className?: string;
-  open: boolean;
-  setAddInstructionsOpen: (open: boolean) => void;
+  map: google.maps.Map | null;
+  setAddInfoOpen: (addInfoOpen: boolean) => void;
+  addInfoStep: 'greeting' | 'set-marker' | 'set-radius' | 'set-form';
+  setAddInfoStep: (
+    addInfoStep: 'greeting' | 'set-marker' | 'set-radius' | 'set-form',
+  ) => void;
 }
 
 class AddInstructions extends React.Component<Props, {}> {
+  private addInfo: AddInfo | null = null;
+
+  private completeGreetingStep = () => {
+    const { map } = this.props;
+    if (!map) {
+      return;
+    }
+    if (!this.addInfo) {
+      this.addInfo = {
+        marker: null,
+        circle: null,
+      };
+    }
+    const { setAddInfoStep } = this.props;
+    setAddInfoStep('set-marker');
+  };
+
+  private completeSetMarkerStep = () => {
+    const { map } = this.props;
+    if (!map) {
+      return;
+    }
+    const { setAddInfoStep } = this.props;
+    if (!this.addInfo) {
+      return;
+    }
+    if (this.addInfo.marker) {
+      this.addInfo.marker.setMap(null);
+    }
+    if (this.addInfo.circle) {
+      this.addInfo.circle.setMap(null);
+    }
+
+    const marker = new google.maps.Marker({
+      map,
+      position: map.getCenter(),
+      draggable: true,
+    });
+    const circle = new google.maps.Circle({
+      map,
+      center: map.getCenter(),
+      editable: true,
+      radius: 20000,
+    });
+
+    const meterToKm = (meter: number) => (meter / 1000).toFixed(2);
+    const infoWindow = new google.maps.InfoWindow({
+      content: `${meterToKm(circle.getRadius())}km`,
+    });
+
+    infoWindow.open(map, marker);
+
+    marker.addListener('drag', () => {
+      const position = marker.getPosition();
+      if (!position) {
+        return;
+      }
+      circle.setCenter(new google.maps.LatLng(position.lat(), position.lng()));
+    });
+    marker.addListener('click', () => {
+      if (!map) {
+        return;
+      }
+      infoWindow.open(map, marker);
+    });
+    circle.addListener('center_changed', () => {
+      marker.setPosition(circle.getCenter());
+    });
+    circle.addListener('radius_changed', () => {
+      infoWindow.setContent(`${meterToKm(circle.getRadius())}km`);
+    });
+
+    this.addInfo.marker = marker;
+    this.addInfo.circle = circle;
+    setAddInfoStep('set-radius');
+  };
+
+  private completeSetRadiusStep = () => {
+    const { setAddInfoStep } = this.props;
+    if (!this.addInfo) {
+      setAddInfoStep('set-marker');
+    }
+    setAddInfoStep('set-form');
+  };
+
+  private submit = () => {
+    const { setAddInfoStep } = this.props;
+
+    setAddInfoStep('set-form');
+  };
+
   private close = () => {
-    const { setAddInstructionsOpen } = this.props;
-    setAddInstructionsOpen(false);
+    if (this.addInfo) {
+      if (this.addInfo.marker) {
+        this.addInfo.marker.setMap(null);
+      }
+      if (this.addInfo.circle) {
+        this.addInfo.circle.setMap(null);
+      }
+    }
+    const { setAddInfoOpen, setAddInfoStep } = this.props;
+    setAddInfoOpen(false);
+    setAddInfoStep('greeting');
   };
 
   public render() {
-    const { className, open } = this.props;
+    const { className, addInfoStep, setAddInfoStep } = this.props;
     return (
-      <div className={`${className} ${open ? 'open' : ''}`}>
-        <div className="shadow" onClick={this.close} />
+      <div className={className}>
         <div className="box">
           <div className="top">
             <h2>Add information to this map</h2>
@@ -27,44 +136,83 @@ class AddInstructions extends React.Component<Props, {}> {
               <MdClose size={24} />
             </button>
           </div>
-          <p>
-            We rely on volunteer contributors to keep the information on this
-            map as complete and up-to-date as possible. If you would like to
-            help us by contributing information, either adding new projects or
-            correcting information, there are a couple of different ways you may
-            do so:
-          </p>
-          <ul>
-            <li>
-              <a
-                href="https://docs.google.com/forms/d/e/1FAIpQLSd4rbN_UPk3RW3a_9sIbBmdv6pH30jqzWKv8Hem9YDRr-uAfQ/viewform"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                Fill out our information form
-              </a>
-            </li>
-            <li>
-              <a
-                href="https://github.com/reach4help/reach4help/tree/master/map/src/data"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                Update the information yourself on GitHub
-              </a>
-            </li>
-          </ul>
-          <p>
-            If you have any questions, you can reach us as at&nbsp;
-            <a
-              href="mailto:map@reach4help.org"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              map@reach4help.org
-            </a>
-            .
-          </p>
+          {addInfoStep === 'greeting' && (
+            <>
+              <p>Hi there!</p>
+              <footer>
+                <button
+                  type="button"
+                  className="next-button"
+                  onClick={this.completeGreetingStep}
+                >
+                  Continue
+                </button>
+              </footer>
+            </>
+          )}
+          {addInfoStep === 'set-marker' && (
+            <>
+              <p>Set marker!</p>
+              <footer>
+                <button
+                  type="button"
+                  className="next-button"
+                  onClick={this.completeSetMarkerStep}
+                >
+                  Continue
+                </button>
+                <button
+                  type="button"
+                  className="prev-button"
+                  onClick={() => setAddInfoStep('greeting')}
+                >
+                  Back
+                </button>
+              </footer>
+            </>
+          )}
+          {addInfoStep === 'set-radius' && (
+            <>
+              <p>Hello, World!</p>
+              <footer>
+                <button
+                  type="button"
+                  className="next-button"
+                  onClick={this.completeSetRadiusStep}
+                >
+                  Continue
+                </button>
+                <button
+                  type="button"
+                  className="prev-button"
+                  onClick={() => setAddInfoStep('set-marker')}
+                >
+                  Back
+                </button>
+              </footer>
+            </>
+          )}
+          {addInfoStep === 'set-form' && (
+            <>
+              <p>Hello, World!</p>
+              <footer>
+                <button
+                  type="button"
+                  className="next-button"
+                  onClick={this.submit}
+                >
+                  Submit
+                </button>
+                <button
+                  type="button"
+                  className="prev-button"
+                  onClick={() => setAddInfoStep('set-radius')}
+                >
+                  Back
+                </button>
+              </footer>
+            </>
+          )}
         </div>
       </div>
     );
@@ -73,37 +221,24 @@ class AddInstructions extends React.Component<Props, {}> {
 
 export default styled(AddInstructions)`
   z-index: 1000;
-  display: flex;
-  position: fixed;
+  font-size: 1rem;
+  text-align: left;
+  position: absolute;
   top: 0;
   left: 0;
   bottom: 0;
   right: 0;
-  align-items: center;
-  justify-content: center;
+  display: flex;
+  align-items: flex-start;
+  justify-content: flex-start;
   pointer-events: none;
-  opacity: 0;
-  transition: opacity 0.2s;
-
-  &.open {
-    opacity: 1;
-    pointer-events: all;
-  }
-
-  > .shadow {
-    z-index: 1000;
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    background: rgba(0, 0, 0, 0.5);
-  }
 
   > .box {
     z-index: 1100;
     margin: ${p => p.theme.spacingPx}px;
-    max-width: 600px;
+    margin-top: ${p => p.theme.spacingPx * 5}px;
+    width: 100%;
+    max-width: 470px;
     max-height: 100%;
     max-height: calc(100% - ${p => p.theme.spacingPx * 4}px);
     background: #fff;
@@ -112,6 +247,7 @@ export default styled(AddInstructions)`
     background: #fff;
     padding: ${p => p.theme.spacingPx}px;
     overflow-y: auto;
+    pointer-events: auto;
 
     .top {
       display: flex;
@@ -135,6 +271,20 @@ export default styled(AddInstructions)`
         &:hover {
           color: ${p => p.theme.textColorLight};
         }
+      }
+    }
+
+    footer {
+      display: flex;
+      flex-direction: row-reverse;
+      justify-content: space-between;
+
+      > .next-button {
+        ${buttonPrimary}
+      }
+
+      > .prev-button {
+        ${button}
       }
     }
   }
